## Natural-CORE v0.8.8.0  - Added N.event.isNumberRelatedKeys function that evaluate if the key is related to numeric input.
## Natural-DATA v0.8.2.24
 - N.formatter : Added a fourth argument to specify the datepicker
options in the "date" rule.
 - N.formatter : You can specify a formatting rule such as Ymd as the
second argument of the "date" rule.
 - N.formatter : I changed the execution logic of the "time" rule.

## Natural-UI v0.8.19.15
 - N.grid, N.list : Added a feature that scrolls to the last selected
row when call the select or check function.
 - N.grid, N.list : Fixed a bug where rowHandler was executed when
rowHandlerBeforeBind event handler was defined.
 - N.datepicker : Fixed bugs where non-numeric characters were not
keydowned and the date was not immediately displayed on the first time
input and was not validated.
 - N.datepicker : Added a feature that closes the Datepicker dialog when
the Enter key is pressed on the context element.
 - N.datepicker : Performance improvements.
 - N.datepicker : Appended multilingual messages
 - N.datepicker : Moved the execution point of onBeforeShow event to
before the date element is automatically selected.
 - N.datepicker : Fixed a bug that behaved abnormally when the context
element with the shareEle option applied and the context element without
the shareEle option existed together.
 - N.form : The bind method removes the logic that unbinds the event
before binding it, and changed it to bind only when there are no related
events.

## natural.config.js
 - Added N.datepicker related multilingual messages

## For more information on added and changed features, refer to the API
manual(http://bbalganjjm.github.io/natural_js/)