## Natural-ARCHITECTURE v0.8.2.1  - N.comm.request.param : Fixed a bug where the first 1 character of first parameter name was truncated
## Natural-DATA v0.8.2.29
 - N.formatter : Modify the css-positon related logic of the "date" rule

## Natural-UI v0.8.19.53
 - All components : Fixed a bug where ignore the default value of
options when options's attribute value type is object.
that default value was replaced with the value of N.config or
user-specified option value.
 - All components : Improved consistency with related to applying and
extending options
 - N.grid : Added the misc.fixedcolRootContainer option that can correct
the broken shape of the grid in mobile browsers when exist the element
with dynamically adjust-height on the Grid when fixing a columns by
specifying the "fixedcol" option
 - N.form : Fixed a bug that the focusout.form.format event was
accumulated every time the bind method was called and the value of the
element to which the formatter was applied was not updated.
 - N.form : Removed the code that unbind the keyup.form.dataSync event.

## Natural-UI.Shell v0.8.1.10
 - N.notify : Fixed a bug where ignore the default value of options when
options's attribute value type is object.
that default value was replaced with the value of N.config or
user-specified option value.
 - N.docs : Added a feature that If the second argument of the remove
method is true, the function will close the page unconditionally

## natural.config.js
 - Added N.context.attr("ui").grid.misc.fixedcolRootContainer property
and that's description.