## Natural-ARCHITECTURE v0.8.2.2  - N.cont : Improved the accuracy of logic for the separate the pages when exist the two or more child pages with the same page-id in the parent page.
## Natural-UI v0.8.20.25
 - N.alert : Fixed a bug where scrolling moves to the top if the height
of the popup overs the size of the browser window.
 - N.form : Added the event namespace to the "format" and "unformat"
event of the bind function.
 - N.form : Added event items (format.formatter, unformat.formatter) to
be removed from the unbind method.
 - N.form : Moved the location of this.revertData object to
this.options.revertData.
 - N.form : Added autoUnbind option to automatically call the unbind
method when calling the bind and add methods more than once.
   * Added the this.options.state variable to store the state of what
method was executed first, even if the bind function is called
internally. That property ㅛrelated bind method.
 - N.form : Removed the logic that makes this.options.row value to -1 in
the unbind method.
 - N.form : Removed getRadioNCheckBoxSelectorStr function.
 - N.grid, N.list : Fixed a bug where the addSelect option was not
applied, and changed the default value to false.
 - N.grid, N.list : Added the addScroll option that disable the feature
for automatically scroll to added rows when calling the add method.
 - N.grid, N.list : Added the selectScroll option that disable the
feature for automatically scroll to added rows when calling the select
method.
 - N.grid, N.list : Added the checkScroll option that disable the
feature for automatically scroll to added rows when calling the check
method.
 - N.grid, N.list : Added the validateScroll option that disable the
feature for automatically scroll to rows that exist the input element
that failed validation when calling the validate method.
 - N.grid : Changed the horizontal scroll to always be displayed because
of problems related to overflow(css) and position(css) of mobile
browsers when applying fixedcol option.

## Natural-UI.Shell v0.8.1.20
 - N.docs : Added msgContext option to define the area where the prompt
message is displayed.
 - N.docs : Added addLast option to decide whether to add page tabs
backward or forward.
 - N.docs : Added tabScroll option to scroll tabs left and right.
   * If the tabScroll option is true, the menu selected when clicking a
menu in the menu list will be only selected without going to the front.
Invisible tabs are automatically scrolled to that tab.
 - N.docs : Added entireLoadExcludeURLs option that register exclude
URLs for onBeforeEntireLoad and onEntireLoad events.

## natural.ui.css
 - Modified .tab__>ul.tab_scroll__.effect__ style.
 - Added .docs__>.docs_tab_context__>ul.docs_tabs__.effect__ style.
 - Added .popup__.alert__.block_overlay_msg__ .msg_box__ >
.view_context__ style to prevent style inheritance on iOS.