Bug fixes.
N.alert : Fixed a bug that when N.datepicker and N.validator are used
together, if there is no value, a validation message is displayed and an
error occurs.
 * N.datepicker : Fixed a bug that caused Datepicker to be recreated
even when moving the cursor from side to side in an input element.
 * N.datepicker : Fixed bug where Datepicker did not close when ESC or
TAB key was pressed.
 * N.datepicker : Fixed a bug that caused the Datepicker to break when
entering an incomplete date value.