Rename sitemap.xml and relocate to root directory
Moved sitemap.xml from the etc directory to the root of the project for better accessibility and adherence to standard directory structures. This change simplifies URL references and eases access for web crawlers and other tools. Adjustments in corresponding documentation and scripts may be needed to reflect this path update.